ModelDevelopment/RAG/requirements.txt
# Core ML dependencies
vllm>=0.5.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# MLflow for experiment tracking
mlflow>=2.10.0

# Hyperparameter optimization
optuna>=3.5.0

# Google Cloud
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
google-cloud-monitoring>=2.15.0
google-cloud-pubsub>=2.18.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
scikit-learn>=1.3.0

# Bias detection
fairlearn>=0.9.0

# Utilities
secure-smtplib>=0.1.1
tqdm>=4.66.0